move geocoding methods to geocoder.py
